# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Novate Platform API",
        "description": "Enterprise Digital Workspace Platform API. This API provides endpoints for authentication, organization management, dynamic form building, data collection, project management, analytics, reports, and notifications.",
        "contact": {
            "name": "Novate Platform Support",
            "email": "support@novate.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/submissions": {
            "get": {
                "tags": [
                    "Submissions"
                ],
                "summary": "List submissions",
                "description": "Returns paginated list of submissions for the current organization.",
                "operationId": "listSubmissions",
                "parameters": [
                    {
                        "name": "form_uuid",
                        "in": "query",
                        "description": "Filter by form UUID",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Submissions retrieved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Create submission",
                "description": "Creates a new form submission with payload data. Supports file uploads, GPS coordinates, and offline sync via external_id.",
                "operationId": "createSubmission",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "form_uuid": {
                                        "description": "Form UUID",
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "submitted"
                                    },
                                    "payload": {
                                        "description": "Form field values as key-value pairs",
                                        "type": "object"
                                    },
                                    "external_id": {
                                        "description": "Client-generated ID for offline sync deduplication",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "project_uuid": {
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    },
                                    "device_metadata": {
                                        "description": "Browser/platform info",
                                        "type": "object",
                                        "nullable": true
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "files": {
                                        "description": "File uploads keyed by field_key",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Submission stored"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/submissions/{submission}": {
            "get": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Get submission",
                "description": "Returns a single submission with form, user, and file data.",
                "operationId": "getSubmission",
                "parameters": [
                    {
                        "name": "submission",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Submission retrieved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Update submission",
                "description": "Updates an existing submission. Cannot update already submitted submissions.",
                "operationId": "updateSubmission",
                "parameters": [
                    {
                        "name": "submission",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "payload": {
                                        "type": "object"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "files": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Submission updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Delete submission",
                "description": "Deletes a submission and its associated files.",
                "operationId": "deleteSubmission",
                "parameters": [
                    {
                        "name": "submission",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Submission deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "token",
                "description": "Enter your Bearer token obtained from auth/login or auth/register."
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}